Changed NVME_IOCTL_DEVICE_STATE to accept enum.
Changed NVME_IOCTL_DEVICE_STATE to accept the enumeration directly, rather
than force the interface to creating a structure to place that enum within.
This essentially saves the creation of a structure each time this IOCTL needs
to be called. Skipping a step to make it faster and easier to use.