Remove pin based IRQ support.
Pin based IRQ support, if needed will have to be tested within the dnvme
driver because it doesn't fit well within the user space app design. Because
this is a hybrid approach and certain things couldn't be tested within
user space, it was always the intent to have some things tested in the
dnvme kernel space.