Changed nvme_device to public and private params
The nvme device was a single structure, now refactored to have public
and private members for nvme device, which make it consistent as per
the design and have user access only public portion of this device
structure. Changed across dnvme driver files where the device parameters
are accessed.
Added IOCTL_GET_DEVICE metrics to return public parameters of this device
to the user when requested.