Bugfix: Avoid accidental system freeze.
The NVMe spec states the following:

Creating an I/O Submission Queue that utilizes a PRP List is only valid if the
controller supports non-contiguous queues as indicated in CAP.CQR.

Software determines the maximum number of entries supported per queue (CAP.MQES)
and whether the queues are required to be physically contiguous (CAP.CQR).

Thus we are now checking during kernel mode Q preparation that these conditions
are met. The desire to create discontiguous memory to back an IOQ will only
yield a system with an unknown outcome. This is one of those “shall not” do
statements in the spec. There is no error reporting mechanism available for
this situation and thus should never be allowed. The compliance suite should
allow illegal conditions to be presented to hdw, but not at the expense of a
kernel core dump.