IRQ design and implementation for dnvme
This is a mostly complete IRQ design and implementation for the dnvme driver, it supports
MSI-X, MSI-Single, MSI-Multi and interrupts disabled. The IRQ design supports both ACQ and IOCQ
interrupts. You can associate multiple Q's for the same int vector. The top half will mask
the individual interrupt vector and defers any work to a bottom half using workq's.
The IRQ code is desinged to attached on top of the existing polling scheme and can be disabled
or enabled. The code is tested with MSI-X in QEMU and works for all existing unit tests in this
repo.