Bugfix: Generic mapping of all BAR registers.
BAR0 was assumed to point to the MSIX tables, but according to the spec, BAR4&5
can also do this job. Thus the BAR registers are all now generically mapped
and later dnvme may interpret what they point to. So if MSIX is supported and
tnvme attempts to enable that IRFQ scheme then BAR4&5, i.e. BAR2 in 64 bit
world, will be used to access the MSIX IRQ tables.