Removed dependancy upon nvme_cmds enum.
This enum was used to decide if a cmd was an admin cmd or not from within dnvme.
This however isn't needed since any cmd issued into SQ ID = 0 is thus by
definition an admin cmd. There is no good reason to track this knowledge as
an alternative variable. It is redundant.