Bugfix: a CE's status field was being lopped off.
Certain cmds need special tracking in dnvme, like create IOSQ/IOCQ and
delete IOSQ/IOCQ. Then when they are reaped the status of the
deletion or creation should force dnvme to stop tracking those
resources. This logic was broken because the status field of a CE
is 15 bits and it was being lopped off by passing to functions as
8 bits. Thus, cleanup wasn't being performed and a memory leak
was the result.